+++ Demos/Python/DemoFISTA_artifacts2D.py
@@ -22,7 +22,7 @@ import os
 import tomophantom
 from tomophantom.supp.qualitymetrics import QualityTools
 
-model = 4 # select a model
+model = 12 # select a model
 N_size = 512 # set dimension of the phantom
 # one can specify an exact path to the parameters file
 # path_library2D = '../../../PhantomLibrary/models/Phantom2DLibrary.dat'
@@ -64,7 +64,7 @@ noisy_sino_misalign = _Artifacts_(sinogram = sino_an, \
 noisy_zing_stripe = _Artifacts_(sinogram = sino_an, \
                                   noise_type='Poisson', noise_sigma=10000, noise_seed = 0, \
                                   zingers_percentage=0.25, zingers_modulus = 10,
-                                  stripes_percentage = 1.0, stripes_maxthickness = 1.0)
+                                  stripes_percentage = 2.0, stripes_maxthickness = 1.0)
 
 plt.figure()
 plt.rcParams.update({'font.size': 21})
@@ -73,7 +73,7 @@ plt.colorbar(ticks=[0, 150, 250], orientation='vertical')
 plt.title('{}''{}'.format('Analytical noisy sinogram with artifacts.',model))
 #%%
 from tomobar.methodsDIR import RecToolsDIR
-RectoolsDIR = RecToolsDIR(DetectorsDimH = P,  # DetectorsDimH # detector dimension (horizontal)
+Rectools = RecToolsDIR(DetectorsDimH = P,  # DetectorsDimH # detector dimension (horizontal)
                     DetectorsDimV = None,  # DetectorsDimV # detector dimension (vertical) for 3D case only
                     CenterRotOffset = None, # Center of Rotation (CoR) scalar (for 3D case only)
                     AnglesVec = angles_rad, # array of angles in radians
@@ -83,28 +83,28 @@ RectoolsDIR = RecToolsDIR(DetectorsDimH = P,  # DetectorsDimH # detector dimensi
 print ("%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%")
 print ("Reconstructing analytical sinogram using FBP (tomobar)...")
 print ("%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%")
-FBPrec_ideal = RectoolsDIR.FBP(sino_an)  # ideal reconstruction
-FBPrec_error = RectoolsDIR.FBP(noisy_zing_stripe) # reconstruction with artifacts
-FBPrec_misalign = RectoolsDIR.FBP(noisy_sino_misalign) # reconstruction with misalignment
+FBPrec_ideal = Rectools.FBP(sino_an)  # ideal reconstruction
+FBPrec_error = Rectools.FBP(noisy_zing_stripe) # reconstruction with artifacts
+FBPrec_misalign = Rectools.FBP(noisy_sino_misalign) # reconstruction with misalignment
 
 plt.figure()
 plt.subplot(131)
-plt.imshow(FBPrec_ideal, vmin=0, vmax=1, cmap="gray")
-plt.colorbar(ticks=[0, 0.5, 1], orientation='vertical')
+plt.imshow(FBPrec_ideal, vmin=0, vmax=3, cmap="gray")
+plt.colorbar(ticks=[0, 0.5, 3], orientation='vertical')
 plt.title('Ideal FBP reconstruction')
 plt.subplot(132)
-plt.imshow(FBPrec_error, vmin=0, vmax=1, cmap="gray")
-plt.colorbar(ticks=[0, 0.5, 1], orientation='vertical')
+plt.imshow(FBPrec_error, vmin=0, vmax=3, cmap="gray")
+plt.colorbar(ticks=[0, 0.5, 3], orientation='vertical')
 plt.title('Erroneous data FBP Reconstruction')
 plt.subplot(133)
-plt.imshow(FBPrec_misalign, vmin=0, vmax=1, cmap="gray")
-plt.colorbar(ticks=[0, 0.5, 1], orientation='vertical')
+plt.imshow(FBPrec_misalign, vmin=0, vmax=3, cmap="gray")
+plt.colorbar(ticks=[0, 0.5, 3], orientation='vertical')
 plt.title('Misaligned noisy FBP Reconstruction')
 plt.show()
 
 plt.figure() 
-plt.imshow(abs(FBPrec_ideal-FBPrec_error), vmin=0, vmax=1, cmap="gray")
-plt.colorbar(ticks=[0, 0.5, 1], orientation='vertical')
+plt.imshow(abs(FBPrec_ideal-FBPrec_error), vmin=0, vmax=3, cmap="gray")
+plt.colorbar(ticks=[0, 0.5, 3], orientation='vertical')
 plt.title('FBP reconsrtuction differences')
 #%%
 print ("%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%")
@@ -131,8 +131,8 @@ RecFISTA_LS_reg = RectoolsIR.FISTA(noisy_zing_stripe,
                                    regularisation_parameter = 0.003,
                                    lipschitz_const = lc)
 plt.figure()
-plt.imshow(RecFISTA_LS_reg, vmin=0, vmax=1, cmap="gray")
-plt.colorbar(ticks=[0, 0.5, 1], orientation='vertical')
+plt.imshow(RecFISTA_LS_reg, vmin=0, vmax=3, cmap="gray")
+plt.colorbar(ticks=[0, 0.5, 3], orientation='vertical')
 plt.title('FISTA-LS-TV reconstruction')
 #%%
 RectoolsIR = RecToolsIR(DetectorsDimH = P,  # DetectorsDimH # detector dimension (horizontal)
@@ -147,7 +147,7 @@ RectoolsIR = RecToolsIR(DetectorsDimH = P,  # DetectorsDimH # detector dimension
                     device='gpu')
 
 RecFISTA_Huber_reg = RectoolsIR.FISTA(noisy_zing_stripe, 
-                                   huber_data_threshold = 3.0,
+                                   huber_data_threshold = 3.5,
                                    iterationsFISTA = 350, 
                                    regularisation = 'ROF_TV', 
                                    regularisation_parameter = 0.003,
@@ -156,11 +156,11 @@ RecFISTA_Huber_reg = RectoolsIR.FISTA(noisy_zing_stripe,
 plt.figure()
 plt.subplot(121)
 plt.imshow(RecFISTA_LS_reg, vmin=0, vmax=3, cmap="gray")
-plt.colorbar(ticks=[0, 0.5, 1], orientation='vertical')
+plt.colorbar(ticks=[0, 0.5, 3], orientation='vertical')
 plt.title('FISTA-LS-TV reconstruction')
 plt.subplot(122)
 plt.imshow(RecFISTA_Huber_reg, vmin=0, vmax=3, cmap="gray")
-plt.colorbar(ticks=[0, 0.5, 1], orientation='vertical')
+plt.colorbar(ticks=[0, 0.5, 3], orientation='vertical')
 plt.title('FISTA-Huber-TV reconstruction')
 plt.show()
 
@@ -175,7 +175,6 @@ print("RMSE for FISTA-Huber-TV is {}".format(RMSE_FISTA_HUBER_TV))
 print ("%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%")
 print ("Reconstructing using FISTA-OS method (tomobar)")
 print ("%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%")
-
 from tomobar.methodsIR import RecToolsIR
 RectoolsIR = RecToolsIR(DetectorsDimH = P,  # DetectorsDimH # detector dimension (horizontal)
                     DetectorsDimV = None,  # DetectorsDimV # detector dimension (vertical) for 3D case only
@@ -193,7 +192,7 @@ lc = RectoolsIR.powermethod() # calculate Lipschitz constant
 
 # Run FISTA reconstrucion algorithm with regularisation 
 RecFISTA_LS_reg = RectoolsIR.FISTA(noisy_zing_stripe,
-                                   iterationsFISTA = 15, 
+                                   iterationsFISTA = 20, 
                                    regularisation = 'ROF_TV', 
                                    regularisation_parameter = 0.01,
                                    regularisation_iterations = 300,
@@ -211,8 +210,8 @@ RectoolsIR = RecToolsIR(DetectorsDimH = P,  # DetectorsDimH # detector dimension
                     device='gpu')
 
 RecFISTA_Huber_reg = RectoolsIR.FISTA(noisy_zing_stripe, 
-                                   huber_data_threshold=5,
-                                   iterationsFISTA = 15, 
+                                   huber_data_threshold=4.0,
+                                   iterationsFISTA = 20, 
                                    regularisation = 'ROF_TV', 
                                    regularisation_parameter = 0.01,
                                    regularisation_iterations = 300,
@@ -238,6 +237,86 @@ RMSE_FISTA_HUBER_TV = Qtools.rmse()
 print("RMSE for FISTA-OS-LS-TV reconstruction is {}".format(RMSE_FISTA_LS_TV))
 print("RMSE for FISTA-OS-Huber-TV is {}".format(RMSE_FISTA_HUBER_TV))
 
+#%%
+RectoolsIR = RecToolsIR(DetectorsDimH = P,  # DetectorsDimH # detector dimension (horizontal)
+                    DetectorsDimV = None,  # DetectorsDimV # detector dimension (vertical) for 3D case only
+                    CenterRotOffset = None, # Center of Rotation (CoR) scalar (for 3D case only)
+                    AnglesVec = angles_rad, # array of angles in radians
+                    ObjSize = P, # a scalar to define reconstructed object dimensions
+                    datafidelity='LS', #data fidelity, choose LS, PWLS, GH (wip), Student (wip)
+                    nonnegativity='ENABLE', # enable nonnegativity constraint (set to 'ENABLE')
+                    OS_number = 12, # the number of subsets, NONE/(or > 1) ~ classical / ordered subsets
+                    tolerance = 1e-06, # tolerance to stop outer iterations earlier
+                    device='gpu')
+
+RecFISTA_Huber_RingModel = RectoolsIR.FISTA(noisy_zing_stripe, 
+                                   huber_data_threshold=4.0,
+                                   ring_model_horiz_size = 9,
+                                   ring_model_vert_size = 1,
+                                   iterationsFISTA = 20, 
+                                   regularisation = 'ROF_TV', 
+                                   regularisation_parameter = 0.01,
+                                   regularisation_iterations = 300,
+                                   lipschitz_const = lc)
+
+plt.figure()
+plt.subplot(121)
+plt.imshow(RecFISTA_Huber_reg, vmin=0, vmax=3, cmap="gray")
+plt.colorbar(ticks=[0, 0.5, 1], orientation='vertical')
+plt.title('FISTA-OS-Huber-TV reconstruction')
+plt.subplot(122)
+plt.imshow(RecFISTA_Huber_RingModel, vmin=0, vmax=3, cmap="gray")
+plt.colorbar(ticks=[0, 0.5, 1], orientation='vertical')
+plt.title('FISTA-OS-Huber-RingModel-TV reconstruction')
+plt.show()
+
+# calculate errors 
+Qtools = QualityTools(phantom_2D, RecFISTA_Huber_reg)
+RMSE_FISTA_HUBER_TV = Qtools.rmse()
+Qtools = QualityTools(phantom_2D, RecFISTA_Huber_RingModel)
+RMSE_FISTA_HUBER_RING_TV = Qtools.rmse()
+
+print("RMSE for FISTA-OS-Huber-TV is {}".format(RMSE_FISTA_HUBER_TV))
+print("RMSE for FISTA-OS-Huber-RingModel-TV is {}".format(RMSE_FISTA_HUBER_RING_TV))
+#%%
+RectoolsIR = RecToolsIR(DetectorsDimH = P,  # DetectorsDimH # detector dimension (horizontal)
+                    DetectorsDimV = None,  # DetectorsDimV # detector dimension (vertical) for 3D case only
+                    CenterRotOffset = None, # Center of Rotation (CoR) scalar (for 3D case only)
+                    AnglesVec = angles_rad, # array of angles in radians
+                    ObjSize = P, # a scalar to define reconstructed object dimensions
+                    datafidelity='LS', #data fidelity, choose LS, PWLS, GH (wip), Student (wip)
+                    nonnegativity='ENABLE', # enable nonnegativity constraint (set to 'ENABLE')
+                    OS_number = 12, # the number of subsets, NONE/(or > 1) ~ classical / ordered subsets
+                    tolerance = 1e-10, # tolerance to stop outer iterations earlier
+                    device='gpu')
+
+RecFISTA_Sudentst = RectoolsIR.FISTA(noisy_zing_stripe, 
+                                   student_data_threshold = 4.5,
+                                   iterationsFISTA = 150, 
+                                   regularisation = 'ROF_TV', 
+                                   regularisation_parameter = 0.0008,
+                                   regularisation_iterations = 350,
+                                   lipschitz_const = lc)
+
+plt.figure()
+plt.subplot(121)
+plt.imshow(RecFISTA_Huber_reg, vmin=0, vmax=3, cmap="gray")
+plt.colorbar(ticks=[0, 0.5, 1], orientation='vertical')
+plt.title('FISTA-OS-Huber-TV reconstruction')
+plt.subplot(122)
+plt.imshow(RecFISTA_Sudentst, vmin=0, vmax=3, cmap="gray")
+plt.colorbar(ticks=[0, 0.5, 1], orientation='vertical')
+plt.title('FISTA-OS-Studentst-TV reconstruction')
+plt.show()
+
+# calculate errors 
+Qtools = QualityTools(phantom_2D, RecFISTA_Huber_reg)
+RMSE_FISTA_HUBER_TV = Qtools.rmse()
+Qtools = QualityTools(phantom_2D, RecFISTA_Sudentst)
+RMSE_FISTA_STUDENT_TV = Qtools.rmse()
+
+print("RMSE for FISTA-OS-Huber-TV is {}".format(RMSE_FISTA_HUBER_TV))
+print("RMSE for FISTA-OS-Studentst-TV is {}".format(RMSE_FISTA_STUDENT_TV))
 #%%
 print ("%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%")
 print ("Reconstructing using FISTA-Group-Huber method (tomobar)")