+++ src/Python/conda-recipe/build.sh
@@ -1,5 +1,0 @@
-mkdir ${SRC_DIR}/ToMoBAR
-cp -r "${RECIPE_DIR}/../../../" ${SRC_DIR}/ToMoBAR
-
-cd ${SRC_DIR}/ToMoBAR/src/Python
-$PYTHON setup.py install